chore(serials): split useSerials.ts into focused hook files
useSerials.ts (450 lines) was over the repo's 330-line max-lines lint
limit. Split into serials/query-keys, use-series-queries,
use-series-mutations, use-season-episode-interactions, and
use-season-episode-reviews, grouped by what each hook operates on.
useSerials.ts is now a barrel re-exporting all of them, so every
existing import path (9 consumer files) is unchanged. No hook logic
changed.
